# contracts/openapi/public-api.yaml
openapi: 3.0.3
info:
  title: SmartCashFlow Public API
  version: 0.1.0
  description: Initial contract map for the current MVP API.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
...
paths:
...
  /v1/auth/signup:
    post:
      tags: [identity]
      summary: Create a user account.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
  /v1/auth/login:
    post:
      tags: [identity]
      summary: Authenticate with email and password.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: API session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
  /v1/dashboard/summary:
    get:
      tags: [dashboard]
      summary: Financial summary for a period.
      parameters:
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Dashboard summary.
...
  /v1/dashboard/daily-cashflow:
    get:
      tags: [dashboard]
      summary: Daily cashflow series.
      parameters:
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Daily cashflow.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DateFrom:
      name: date_from
      in: query
      schema:
        type: string
        format: date
    DateTo:
      name: date_to
      in: query
      schema:
        type: string
        format: date
...
  schemas:
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        display_name:
          type: string
          nullable: true
    SignupResponse:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
        display_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
          nullable: true
        user_id:
          type: string
        email:
          type: string
        display_name:
          type: string
          nullable: true
...